+++ perl/LinkedSpec/BootstrapSpec.pm
@@ -16,17 +16,6 @@ use LinkedSpec::OwnerDispatch ();
 
 my $CACHED_BOOTSTRAP_STATE;
 
-#------------------------------------------------------------------------------
-# Function: _require_pkg
-# Purpose : Lazy-load one owner package through the shared dispatch helper.
-# Args    : ($pkg)
-# Returns : true on successful require
-#------------------------------------------------------------------------------
-sub _require_pkg {
- my ($pkg) = @_;
- return LinkedSpec::OwnerDispatch::require_pkg(__PACKAGE__, $pkg)
-}
-
 sub _require_pkg_cb {
  my ($pkg, $name) = @_;
  return LinkedSpec::OwnerDispatch::require_pkg_cb(__PACKAGE__, $pkg, $name)